Feat(file_upload): Create an handler that can program a board via POST
The POST request needs a boad field, a port field and a sketch field
which is a file.

It returns an error if one of these is missing, or proceeds to save the
sketch and program the board.

The results of the board programming are sent via websocket as usual.